project1/Effective_Mobile_AQA/requirements.txt
selenium==4.17.2
pytest==7.4.4
allure-pytest==2.13.2
webdriver-manager==4.0.1
